[REF] Tests: move LineBreak tests to LineBreak plugin
