+++ packages/core/test/VDocument.test.ts
@@ -8,7 +8,6 @@ import { BasicEditor } from '../../../bundles/BasicEditor';
 const deleteForward = (editor: JWEditor): void => editor.execCommand('deleteForward');
 const deleteBackward = (editor: JWEditor): void => editor.execCommand('deleteBackward');
 const insertParagraphBreak = (editor: JWEditor): void => editor.execCommand('insertParagraphBreak');
-const insertLineBreak = (editor: JWEditor): void => editor.execCommand('insertLineBreak');
 
 describe('VDocument', () => {
     // Note: implementing a test for deleteForward, make sure to implement
@@ -1307,322 +1306,6 @@ describe('VDocument', () => {
                     contentBefore: '<p>]abcd[</p>',
                     stepFunction: insertParagraphBreak,
                     contentAfter: '<p><br></p><p>[]<br></p>',
-                });
-            });
-        });
-    });
-    describe('insertLineBreak', () => {
-        describe('Selection collapsed', () => {
-            describe('Basic', () => {
-                it('should insert a <br> into an empty paragraph', async () => {
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p>[]<br></p>',
-                        stepFunction: insertLineBreak,
-                        contentAfter: '<p><br>[]<br></p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p>[<br>]</p>',
-                        stepFunction: insertLineBreak,
-                        contentAfter: '<p><br>[]<br></p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p><br>[]</p>',
-                        stepFunction: insertLineBreak,
-                        contentAfter: '<p><br>[]<br></p>',
-                    });
-                });
-                it('should insert a <br> at the beggining of a paragraph', async () => {
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p>[]abc</p>',
-                        stepFunction: insertLineBreak,
-                        contentAfter: '<p><br>[]abc</p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p>[] abc</p>',
-                        stepFunction: insertLineBreak,
-                        // The space should have been parsed away.
-                        contentAfter: '<p><br>[]abc</p>',
-                    });
-                });
-                it('should insert a <br> within text', async () => {
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p>ab[]cd</p>',
-                        stepFunction: insertLineBreak,
-                        contentAfter: '<p>ab<br>[]cd</p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p>ab []cd</p>',
-                        stepFunction: insertLineBreak,
-                        // The space is converted to a non-breaking
-                        // space so it is visible (because it's before a
-                        // <br>).
-                        contentAfter: '<p>ab&nbsp;<br>[]cd</p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p>ab[] cd</p>',
-                        stepFunction: insertLineBreak,
-                        // The space is converted to a non-breaking
-                        // space so it is visible (because it's after a
-                        // <br>).
-                        contentAfter: '<p>ab<br>[]&nbsp;cd</p>',
-                    });
-                });
-                it('should insert a line break (2 <br>) at the end of a paragraph', async () => {
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p>abc[]</p>',
-                        stepFunction: insertLineBreak,
-                        // The second <br> is needed to make the first
-                        // one visible.
-                        contentAfter: '<p>abc<br>[]<br></p>',
-                    });
-                });
-            });
-            describe('Consecutive', () => {
-                it('should insert two <br> at the beggining of an empty paragraph', async () => {
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p>[]<br></p>',
-                        stepFunction: (editor: JWEditor) => {
-                            insertLineBreak(editor);
-                            insertLineBreak(editor);
-                        },
-                        contentAfter: '<p><br><br>[]<br></p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p>[<br>]</p>',
-                        stepFunction: (editor: JWEditor) => {
-                            insertLineBreak(editor);
-                            insertLineBreak(editor);
-                        },
-                        contentAfter: '<p><br><br>[]<br></p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p><br>[]</p>',
-                        stepFunction: (editor: JWEditor) => {
-                            insertLineBreak(editor);
-                            insertLineBreak(editor);
-                        },
-                        contentAfter: '<p><br><br>[]<br></p>',
-                    });
-                });
-                it('should insert two <br> at the beggining of a paragraph', async () => {
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p>[]abc</p>',
-                        stepFunction: (editor: JWEditor) => {
-                            insertLineBreak(editor);
-                            insertLineBreak(editor);
-                        },
-                        contentAfter: '<p><br><br>[]abc</p>',
-                    });
-                });
-                it('should insert two <br> within text', async () => {
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p>ab[]cd</p>',
-                        stepFunction: (editor: JWEditor) => {
-                            insertLineBreak(editor);
-                            insertLineBreak(editor);
-                        },
-                        contentAfter: '<p>ab<br><br>[]cd</p>',
-                    });
-                });
-                it('should insert two line breaks (3 <br>) at the end of a paragraph', async () => {
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p>abc[]</p>',
-                        stepFunction: (editor: JWEditor) => {
-                            insertLineBreak(editor);
-                            insertLineBreak(editor);
-                        },
-                        // the last <br> is needed to make the first one
-                        // visible.
-                        contentAfter: '<p>abc<br><br>[]<br></p>',
-                    });
-                });
-            });
-            describe('Format', () => {
-                it('should insert a <br> before a format node', async () => {
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p>abc[]<b>def</b></p>',
-                        stepFunction: insertLineBreak,
-                        // That selection is equivalent to []<b>
-                        contentAfter: '<p>abc<br><b>[]def</b></p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        // That selection is equivalent to []<b>
-                        contentBefore: '<p>abc<b>[]def</b></p>',
-                        stepFunction: insertLineBreak,
-                        contentAfter: '<p>abc<br><b>[]def</b></p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p>abc <b>[]def</b></p>',
-                        stepFunction: insertLineBreak,
-                        // The space is converted to a non-breaking
-                        // space so it is visible (because it's after a
-                        // <br>).
-                        contentAfter: '<p>abc&nbsp;<br><b>[]def</b></p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p>abc<b>[] def </b></p>',
-                        stepFunction: insertLineBreak,
-                        // The space is converted to a non-breaking
-                        // space so it is visible (because it's before a
-                        // <br>).
-                        contentAfter: '<p>abc<br><b>[]&nbsp;def</b></p>',
-                    });
-                });
-                it('should insert a <br> after a format node', async () => {
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p><b>abc</b>[]def</p>',
-                        stepFunction: insertLineBreak,
-                        contentAfter: '<p><b>abc</b><br>[]def</p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        // That selection is equivalent to </b>[]
-                        contentBefore: '<p><b>abc[]</b>def</p>',
-                        stepFunction: insertLineBreak,
-                        contentAfter: '<p><b>abc</b><br>[]def</p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p><b>abc[]</b> def</p>',
-                        stepFunction: insertLineBreak,
-                        // The space is converted to a non-breaking
-                        // space so it is visible (because it's after a
-                        // <br>).
-                        contentAfter: '<p><b>abc</b><br>[]&nbsp;def</p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p><b>abc []</b>def</p>',
-                        stepFunction: insertLineBreak,
-                        // The space is converted to a non-breaking
-                        // space so it is visible (because it's before a
-                        // <br>).
-                        contentAfter: '<p><b>abc&nbsp;</b><br>[]def</p>',
-                    });
-                });
-                it('should insert a <br> at the beginning of a format node', async () => {
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p>[]<b>abc</b></p>',
-                        stepFunction: insertLineBreak,
-                        contentAfter: '<p><br><b>[]abc</b></p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        // That selection is equivalent to []<b>
-                        contentBefore: '<p><b>[]abc</b></p>',
-                        stepFunction: insertLineBreak,
-                        contentAfter: '<p><br><b>[]abc</b></p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p><b>[] abc</b></p>',
-                        stepFunction: insertLineBreak,
-                        // The space should have been parsed away.
-                        contentAfter: '<p><br><b>[]abc</b></p>',
-                    });
-                });
-                it('should insert a <br> within a format node', async () => {
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p><b>ab[]cd</b></p>',
-                        stepFunction: insertLineBreak,
-                        contentAfter: '<p><b>ab</b><br><b>[]cd</b></p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p><b>ab []cd</b></p>',
-                        stepFunction: insertLineBreak,
-                        // The space is converted to a non-breaking
-                        // space so it is visible.
-                        contentAfter: '<p><b>ab&nbsp;</b><br><b>[]cd</b></p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p><b>ab[] cd</b></p>',
-                        stepFunction: insertLineBreak,
-                        // The space is converted to a non-breaking
-                        // space so it is visible.
-                        contentAfter: '<p><b>ab</b><br><b>[]&nbsp;cd</b></p>',
-                    });
-                });
-                it('should insert a line break (2 <br>) at the end of a format node', async () => {
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p><b>abc</b>[]</p>',
-                        stepFunction: insertLineBreak,
-                        // The second <br> is needed to make the first
-                        // one visible.
-                        contentAfter: '<p><b>abc</b><br>[]<br></p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        // That selection is equivalent to </b>[]
-                        contentBefore: '<p><b>abc[]</b></p>',
-                        stepFunction: insertLineBreak,
-                        // The second <br> is needed to make the first
-                        // one visible.
-                        contentAfter: '<p><b>abc</b><br>[]<br></p>',
-                    });
-                    await testEditor(BasicEditor, {
-                        contentBefore: '<p><b>abc[] </b></p>',
-                        stepFunction: insertLineBreak,
-                        // The space should have been parsed away.
-                        // The second <br> is needed to make the first
-                        // one visible.
-                        contentAfter: '<p><b>abc</b><br>[]<br></p>',
-                    });
-                });
-            });
-        });
-        describe('Selection not collapsed', () => {
-            it('should delete the first half of a paragraph, then insert a <br>', async () => {
-                // Forward selection
-                await testEditor(BasicEditor, {
-                    contentBefore: '<p>[ab]cd</p>',
-                    stepFunction: insertLineBreak,
-                    contentAfter: '<p><br>[]cd</p>',
-                });
-                // Backward selection
-                await testEditor(BasicEditor, {
-                    contentBefore: '<p>]ab[cd</p>',
-                    stepFunction: insertLineBreak,
-                    contentAfter: '<p><br>[]cd</p>',
-                });
-            });
-            it('should delete part of a paragraph, then insert a <br>', async () => {
-                // Forward selection
-                await testEditor(BasicEditor, {
-                    contentBefore: '<p>a[bc]d</p>',
-                    stepFunction: insertLineBreak,
-                    contentAfter: '<p>a<br>[]d</p>',
-                });
-                // Backward selection
-                await testEditor(BasicEditor, {
-                    contentBefore: '<p>a]bc[d</p>',
-                    stepFunction: insertLineBreak,
-                    contentAfter: '<p>a<br>[]d</p>',
-                });
-            });
-            it('should delete the last half of a paragraph, then insert a line break (2 <br>)', async () => {
-                // Forward selection
-                await testEditor(BasicEditor, {
-                    contentBefore: '<p>ab[cd]</p>',
-                    stepFunction: insertLineBreak,
-                    // the second <br> is needed to make the first one
-                    // visible.
-                    contentAfter: '<p>ab<br>[]<br></p>',
-                });
-                // Backward selection
-                await testEditor(BasicEditor, {
-                    contentBefore: '<p>ab]cd[</p>',
-                    stepFunction: insertLineBreak,
-                    // the second <br> is needed to make the first one
-                    // visible.
-                    contentAfter: '<p>ab<br>[]<br></p>',
-                });
-            });
-            it('should delete all contents of a paragraph, then insert a line break', async () => {
-                // Forward selection
-                await testEditor(BasicEditor, {
-                    contentBefore: '<p>[abcd]</p>',
-                    stepFunction: insertLineBreak,
-                    contentAfter: '<p><br>[]<br></p>',
-                });
-                // Backward selection
-                await testEditor(BasicEditor, {
-                    contentBefore: '<p>]abcd[</p>',
-                    stepFunction: insertLineBreak,
-                    contentAfter: '<p><br>[]<br></p>',
                 });
             });
         });
